+++ Installation/CHANGES.md
@@ -30,8 +30,12 @@ Release date: June 2020
 ### 3D Triangulations
 -   The free function `CGAL::file_input()` and the member function `CGAL::Triangulation_3::file_input()` 
     have been added. The first allows to load a `Triangulation_3` from an input stream, using functors to create vertices and cells.
-    The second is simply the member function version of the first one. 
+    The second is simply the member function version of the first one.
 
+### 3D Triangulation Data Structure
+-   The free function `CGAL::file_input()` and the member function `CGAL::TDS_3::file_input()` 
+    have been added. The first allows to load a `TDS_3` from an input stream, using functors to create vertices and cells.
+    The second is simply the member function version of the first one. 
 
 ### dD Spatial Searching